requirements.txt
beautifulsoup4==4.12.2
python-telegram-bot==13.13
Requests==2.31.0
lxml==4.9.3
